refactor: reorganize service interfaces; and update package structure
