+++ src/main/java/com/shintadev/shop_dev_app/controller/ProductController.java
@@ -18,7 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.shintadev.shop_dev_app.domain.dto.request.ProductRequest;
 import com.shintadev.shop_dev_app.domain.dto.response.ProductResponse;
-import com.shintadev.shop_dev_app.service.ProductService;
+import com.shintadev.shop_dev_app.service.product.ProductService;
 
 import lombok.RequiredArgsConstructor;
