--- src/main/java/com/shintadev/shop_dev_app/service/OrderService.java
@@ -1,12 +1,0 @@
-package com.shintadev.shop_dev_app.service;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-import com.shintadev.shop_dev_app.base.BaseService;
-
-public interface OrderService {
-
-  // Page<OrderDto> findByUserId(Long userId, Pageable pageable);
-
-}
