--- src/main/java/com/shintadev/shop_dev_app/service/PaymentService.java
@@ -1,5 +1,0 @@
-package com.shintadev.shop_dev_app.service;
-
-public interface PaymentService {
-
-}
