+++ src/main/java/com/shintadev/shop_dev_app/service/order/OrderService.java
@@ -1,0 +1,7 @@
+package com.shintadev.shop_dev_app.service.order;
+
+public interface OrderService {
+
+  // Page<OrderDto> findByUserId(Long userId, Pageable pageable);
+
+}
