+++ src/main/java/com/shintadev/shop_dev_app/service/order/PaymentService.java
@@ -1,0 +1,5 @@
+package com.shintadev.shop_dev_app.service.order;
+
+public interface PaymentService {
+
+}
