+++ src/main/java/com/shintadev/shop_dev_app/service/other/KafkaService.java
@@ -1,4 +1,4 @@
-package com.shintadev.shop_dev_app.service;
+package com.shintadev.shop_dev_app.service.other;
 
 public interface KafkaService {
