+++ src/main/resources/application.yml
@@ -8,7 +8,7 @@ spring:
     import:
       - optional:file:.env[.properties]
   datasource:
-    url: "jdbc:${DB_TYPE}://${DB_HOST}:${DB_PORT}/${DB_NAME}?useSSL=false"
+    url: "jdbc:${DB_TYPE}://${DB_HOST}:${DB_PORT}/${DB_NAME}?useSSL=false&allowPublicKeyRetrieval=true"
     username: ${DB_USERNAME}
     password: ${DB_PASSWORD}
     driver-class-name: com.mysql.cj.jdbc.Driver
